feat: users  타입 분리
